requirements.txt
requests==2.24.0
kafka-python==2.0.1